refactor - centralize axis format descriptions
