+++ src/traces/histogram/attributes.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var barAttrs = require('../bar/attributes');
-var axisHoverFormat = require('../../plots/hoverformat_attributes');
+var axisHoverFormat = require('../../plots/cartesian/axis_format').axisHoverFormat;
 var hovertemplateAttrs = require('../../plots/template_attributes').hovertemplateAttrs;
 var makeBinAttrs = require('./bin_attributes');
 var constants = require('./constants');
