+++ src/traces/histogram2dcontour/attributes.js
@@ -3,7 +3,7 @@
 var histogram2dAttrs = require('../histogram2d/attributes');
 var contourAttrs = require('../contour/attributes');
 var colorScaleAttrs = require('../../components/colorscale/attributes');
-var axisHoverFormat = require('../../plots/hoverformat_attributes');
+var axisHoverFormat = require('../../plots/cartesian/axis_format').axisHoverFormat;
 
 var extendFlat = require('../../lib/extend').extendFlat;
